Adding a router and navbar
