+++ Team4of5/src/Team4of5_App/ProjectManagement/ProjectManagement.js
@@ -6,7 +6,7 @@ class ProjectManagement extends React.Component {
     constructor(props) {
         super(props);
     }
-    
+
     render() {
         return (
             <h1>Project Management</h1>
