+++ Team4of5/src/Team4of5_App/RouteMap.js
@@ -1,0 +1,46 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import BugTracking from './BugTracking/BugTracking.js';
+import Chat from './Chat/Chat.js';
+import ProjectManagement from './ProjectManagement/ProjectManagement.js';
+import Setting from './Setting/Setting.js';
+import Navbar from './Navbar/Nav.js';
+import Menu from './Menu.js';
+
+
+import {
+    BrowserRouter as Router,
+    Route,
+    Link,
+    Redirect,
+    Switch,
+    withRouter
+} from 'react-router-dom';
+
+import PropTypes from "prop-types";
+
+class RouteMap extends React.Component {
+  render() {
+    return(
+      <Router>
+        <div className = 'container'>
+          <Navbar/>
+          <Switch>
+           <Route exact path='/setting' component={Setting} />
+           <Route exact path='/bugtrack' component={BugTracking} />
+           <Route exact path='/chat' component={Chat} />
+            <Route exact path='/menu' component={Menu} />
+           <Route path='/projectmanage' component={ProjectManagement} />
+
+           <Route render={function () {
+              return <p>Not Found</p>
+            }} />
+          </Switch>
+        </div>
+      </Router>
+
+    )
+
+  }
+}
+export default RouteMap;
