+++ Team4of5/src/Team4of5_App/UserLoginSignup/UserLoginSignup.js
@@ -1,0 +1,112 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import registerServiceWorker from '/Users/yd/React/Project/Project/Team4of5/src/index.js';
+import '/Users/yd/React/Project/Project/Team4of5/src/index.css';
+import * as Users from '/Users/yd/React/Project/Project/Team4of5/src/Team4of5_Service/Users.js';
+
+
+class UserLoginSignup extends React.Component {
+
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      email: '',
+      password: '',
+      formBtnTxt: 'Login',
+      redirectToMenu: false
+    };
+    this.handleSubmit = this.handleSubmit.bind(this);
+    this.switchLoginSignup = this.switchLoginSignup.bind(this);
+  }
+
+
+  handleChange(name, event) {
+    let items = this.state;
+    items[name] = event.target.value;
+    this.setState(items);
+  }
+
+  handleSubmit(event) {
+    if (document.getElementById("submitBtn").value == "Login") {
+
+      Users.sign_in_user(this.state.email, this.state.password)
+      .then((User) => {
+        console.log(User);
+        alert("Login Succeed!!");
+          console.log('User Confirm!!');
+          this.setState({redirectToMenu: true});
+
+      }).catch((error) => {
+        console.log(error);
+        alert(error.message);
+      });
+    } else {
+      Users.create_user(this.state.email, this.state.password)
+        .then((User) => {
+          console.log(User);
+          alert('Sign Up Succeed!!');
+        })
+        .catch((error) => {
+          console.log(error);
+          alert(error.message);
+        });
+    }
+    event.preventDefault();
+
+  }
+
+  switchLoginSignup(event) {
+    let submitBtn = document.getElementById("submitBtn");
+    if (submitBtn.value == "Login"){
+      this.state.formBtnTxt = "SignUp";
+      submitBtn.value = "SignUp";
+    } else{
+      submitBtn.value = "Login";
+      this.state.formBtnTxt = "Login";
+    }
+  }
+
+  render() {
+    const { from } = { from: { pathname: '/Menu' } }
+    const { redirectToMenu } = this.state
+
+    // if (redirectToMenu) {
+    //   console.log(from);
+    //   return (
+    //     <Redirect to="/Menu" push />
+    //     //<Redirect to={from}/>
+    //   )
+    // }
+
+    return (
+      <form onSubmit={this.handleSubmit}>
+
+        <button type="button"
+          id="switchBtn"
+          onClick={this.switchLoginSignup}>Login | SignUp</button>
+
+        <div>
+          <label>
+            <div>
+              Email:
+                    </div>
+            <input type="text" value={this.state.email} onChange={this.handleChange.bind(this, 'email')} />
+          </label>
+        </div>
+        <div>
+          <label>
+            <div>
+              Password(at least 6 digits):
+                    </div>
+            <input type="text" value={this.state.first_name} onChange={this.handleChange.bind(this, 'password')} />
+          </label>
+        </div>
+        <input type="submit"
+          id="submitBtn"
+          value={this.state.formBtnTxt} />
+      </form>
+    );
+  }
+}
+export default UserLoginSignup
